backend/requirements.txt
# Core framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Email validation
email-validator>=2.0.0

# Redis for rate limiting
redis>=5.0.0

# HTTP client
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9
